requirements.txt
pandas
numpy
keras
tensorflow
scikit-learn
matplotlib
pillow
jupyter
streamlit